Se ajusta el llamado del constructor de RTGMGateway

+++ ModuloCredCob/Main.vb
@@ -165,7 +165,7 @@ Module Main
     'Código del empleado utilizado para identificar abonos externos
     Public GLOBAL_EmpleadoAbonoExterno As Integer
 
-    Public GLOBAL_Modulo As Short = 4
+    Public GLOBAL_Modulo As Byte = 4
     Public GLOBAL_Empresa As Short
 
 #End Region
